requirements.txt
pyserial==3.5
python-can==4.3.1
numpy==1.25.2
pyqt5==5.15.9